Move `insert_header()` from mock to test_utils
Allows the benchmarking code to use this without having to pull it in from the mock.
